Moved mapPipette into mapBlueprint.lua, restoreGrid into saveGrid.lua, and restoreBurner into saveBurner.lua.  For better portability and readability.
