+++ util/saveGrid.lua
@@ -1,8 +1,9 @@
 --[[ Copyright (c) 2019 robot256 (MIT License)
  * Project: Multiple Unit Train Control
  * File: saveGrid.lua
- * Description: Saves the contents of an equipment grid.
+ * Description: Saves and restores the contents of an equipment grid.
  *    Includes energy, shield, and burner equipment properties.
+ *    Depends on saveBurner.lua.
 --]]
 
 
@@ -17,6 +18,21 @@ function saveGrid(grid)
 		return gridContents
 	else
 		return nil
+	end
+end
+
+function restoreGrid(grid,savedItems)
+	for _,v in pairs(savedItems) do
+		local e = grid.put(v.item)
+		if v.energy then
+			e.energy = v.energy
+		end
+		if v.shield and v.shield > 0 then
+			e.shield = v.shield
+		end
+		if v.burner then
+			restoreBurner(e.burner,v.burner)
+		end
 	end
 end
